moved parsing argument input/output to crate
