+++ shims/fe2o3-amqp/amqp_large_content_test/src/bin/receiver.rs
@@ -1,0 +1,3 @@
+fn main() {
+    
+}
