+++ shims/fe2o3-amqp/amqp_large_content_test/src/bin/sender.rs
@@ -1,0 +1,3 @@
+fn main() {
+    
+}
