+++ shims/fe2o3-amqp/amqp_types_test/Cargo.toml
@@ -6,12 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-serde = "1"
-serde_json = "1"
+interop_test_common = { version = "0.1.0", path = "../interop_test_common" }
 fe2o3-amqp = "0.0.22"
 anyhow = "1"
-ordered-float = "3"
-hex = "0.4"
-uuid = "1"
 tokio = { version = "1", features = ["net", "macros", "rt-multi-thread", "rt"]}
-base64 = "0.13"
